+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Current
 
+- Add an `addComponents` method to add a list of components at once (#31)[https://github.com/adngdb/entity-system-js/issues/31].
+
 ## 1.5.0 (June 21, 2019)
 
 - Revert `getComponentsData` to return an object (entity => component) ([#29](https://github.com/adngdb/entity-system-js/issues/29)). **[BREAKING]**